Separate KaraokeEditPlayfield
